Added JS to fix hidden nav problem
If you were to hide the nav with the hamburger button then resize the
window past the 665px mark the media query would not work and the nav
would still be hidden. Fixed the issue with a little jquery magic.

Also changed the media query trigger mark from 660 to 665 to provide a
smoother transition